Move scroll motion hooks
