+++ components/landing/landing-cta.tsx
@@ -11,7 +11,7 @@ import {
   Zap,
 } from 'lucide-react'
 import { cn } from '@/lib/utils'
-import { useReveal } from '@/lib/landing/use-scroll-motion'
+import { useReveal } from '@/hooks/use-scroll-motion'
 import { useI18n } from '@/lib/i18n/provider'
 
 export function LandingCta() {
